+++ autograph-segment/src/androidMain/kotlin/dev/ynagai/autograph/segment/SegmentTransport.kt
@@ -6,7 +6,9 @@ import com.segment.analytics.kotlin.core.platform.Plugin
 import dev.ynagai.autograph.Envelope
 import dev.ynagai.autograph.EnvelopeSource
 import dev.ynagai.autograph.Transport
+import dev.ynagai.autograph.asJsonObject
 import kotlinx.coroutines.launch
+import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 
 /**
@@ -38,16 +40,16 @@ public class SegmentTransport(
         analytics.add(AutographPlugin(envelopes))
     }
 
-    override fun track(name: String, properties: JsonObject, envelope: Envelope?) {
-        analytics.track(name, properties)
+    override fun track(name: String, properties: Map<String, JsonElement>, envelope: Envelope?) {
+        analytics.track(name, properties.asJsonObject())
     }
 
-    override fun screen(name: String, properties: JsonObject, envelope: Envelope?) {
-        analytics.screen(name, properties)
+    override fun screen(name: String, properties: Map<String, JsonElement>, envelope: Envelope?) {
+        analytics.screen(name, properties.asJsonObject())
     }
 
-    override fun identify(userId: String, traits: JsonObject, envelope: Envelope?) {
-        analytics.identify(userId, traits)
+    override fun identify(userId: String, traits: Map<String, JsonElement>, envelope: Envelope?) {
+        analytics.identify(userId, traits.asJsonObject())
     }
 
     override fun flush() {